+++ game/dictionary.py
@@ -4,6 +4,7 @@ from pyrae import dle
 class DictionaryConnectionError(Exception):
     ...
 
+dle.set_log_level(log_level='CRITICAL')
 
 def validate_word(word):
     search = dle.search_by_word(word=word)